refactor: use `database` instead of `db`
