+++ src/features/auth/repositories/userRepository.ts
@@ -1,4 +1,4 @@
-import db from '@local/libs/db';
+import db from '@local/database/client';
 
 export async function fetchUserByMail(email: string) {
   try {
